pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "geo-analyzer"
version = "0.1.0"
description = "GeoSpatial Site Readiness Analyzer"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "geopandas>=0.14.0",
    "shapely>=2.0.0",
    "pyproj>=3.6.0",
    "geoalchemy2>=0.14.0",
    "h3>=3.7.7",
    "scikit-learn>=1.4.0",
    "weasyprint>=61.0",
    "redis>=5.0.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.6.0",
    "python-multipart>=0.0.9",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "hypothesis>=6.100.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
